保存ボタンクリック時のみ履歴書を上書するように修正

+++ app/controllers/sessions_controller.rb
@@ -3,8 +3,8 @@
 class SessionsController < ApplicationController
   def create
     user = find_or_create_from_auth_hash(auth_hash)
-    log_in(user) if user
-    after_save_for_guest(user) if session['resume_uuid']
+    log_in(user)
+    after_save_for_guest(user)
     redirect_to root_path
   end
 
@@ -20,18 +20,18 @@ class SessionsController < ApplicationController
   end
 
   def find_or_create_from_auth_hash(auth_hash)
-    name = auth_hash['info']['name']
-    email = auth_hash['info']['email']
-    user = User.find_or_create_by(email:)
-    user.update!(name:)
-    user
+    info = auth_hash['info']
+    User.find_or_create_by!(email: info['email']).tap do |user|
+      user.update!(name: info['name'])
+    end
   end
 
   def after_save_for_guest(user)
-    user.skincare_resume.destroy! if user.skincare_resume.present?
+    return unless session['resume_uuid']
+    return unless request.env.dig('omniauth.params', 'button') == 'save'
 
-    resume = SkincareResume.find_by(uuid: session['resume_uuid'])
-    resume.update!(user:)
+    user.skincare_resume&.destroy!
+    SkincareResume.find_by!(uuid: session['resume_uuid']).update!(user:)
     session.delete('resume_uuid')
   end
 end
